netd: fixup arp switch typing, include in device driver header

+++ Applications/netd/lwwire.c
@@ -175,4 +175,4 @@ int device_init( void )
     return 0;
 }
 
-int has_arp = 1;
+uint8_t has_arp = 1;
